+++ cc-backend/src/routes/advertRoutes.ts
@@ -2,6 +2,6 @@ import { Router, Request, Response } from "express";
 import { getItemsPK, getItemsDB, getItemInfo, postConfigToDB } from "../utils";
 import { Error, ItemInfo, Item} from '../types';
 
-const mtRouter = Router();
+const advertRouter = Router();
 
-export default mtRouter;
+export default advertRouter;
