+++ cc-backend/src/routes/systemRoutes.ts
